fix: warn deadlock function in requests
