+++ Cargo.toml
@@ -32,7 +32,6 @@ kameo_macros = { version = "0.15.0", path = "./macros", optional = true }
 
 downcast-rs = "2.0.1"
 dyn-clone = "1.0"
-flume = { version = "0.11.1", features = ["async"] }
 futures = "0.3"
 libp2p = { version = "0.55.0", features = ["cbor", "dns", "kad", "mdns", "macros", "quic", "request-response", "rsa", "serde", "tokio"], optional = true }
 libp2p-identity = { version = "0.2.9", features = ["rand", "rsa"], optional = true }
@@ -40,7 +39,7 @@ linkme = { version= "0.3.28", optional = true }
 once_cell = "1.19"
 rmp-serde = { version = "1.3.0", optional = true }
 serde = { version = "1.0", features = ["derive"] }
-tokio = { version = "1.37", features = ["macros", "rt", "sync", "time"] }
+tokio = { version = "1.44", features = ["macros", "rt", "sync", "time"] }
 tokio-stream = "0.1"
 tracing = { version = "0.1", optional = true }
