+++ src/request/tell.rs
@@ -14,7 +14,7 @@ use crate::{
     Actor, Reply,
 };
 
-use super::{warn_deadlock, WithRequestTimeout, WithoutRequestTimeout};
+use super::{WithRequestTimeout, WithoutRequestTimeout};
 
 /// A request to send a message to an actor without any reply.
 ///
@@ -305,6 +305,34 @@ where
         }
         _ => panic!("unexpected response"),
     }
+}
+
+#[cfg(all(debug_assertions, feature = "tracing"))]
+fn warn_deadlock<A: Actor>(
+    actor_ref: &ActorRef<A>,
+    msg: &'static str,
+    called_at: &'static std::panic::Location<'static>,
+) {
+    use tracing::warn;
+
+    use crate::mailbox::MailboxSender;
+
+    match actor_ref.mailbox_sender() {
+        MailboxSender::Bounded(_) => {
+            if actor_ref.is_current() {
+                warn!("At {called_at}, {msg}");
+            }
+        }
+        MailboxSender::Unbounded(_) => {}
+    }
+}
+
+#[cfg(not(all(debug_assertions, feature = "tracing")))]
+fn warn_deadlock<A: Actor>(
+    _actor_ref: &ActorRef<A>,
+    _msg: &'static str,
+    _called_at: &'static std::panic::Location<'static>,
+) {
 }
 
 #[cfg(test)]
